Change max_size and node_size to be of type size_t.
2007-04-12  Jeffrey Stedfast  <fejj@gnome.org>

	* util/cache.c (cache_new): Change max_size and node_size to be of
	type size_t.

	* gmime/gmime-multipart-encrypted.c
	(g_mime_multipart_encrypted_new): g_object_new() doesn't need a
	second NULL argument.

	* gmime/gmime-utils.c (decode_8bit): Close the iconv descriptor
	and since we are using is_ascii() now, we don't need to use
	unsigned char *'s.


svn path=/trunk/; revision=1122

+++ util/cache.c
@@ -26,7 +26,7 @@
 
 
 Cache *
-cache_new (CacheNodeExpireFunc expire, CacheNodeFreeFunc free_node, unsigned int node_size, unsigned int max_size)
+cache_new (CacheNodeExpireFunc expire, CacheNodeFreeFunc free_node, size_t node_size, size_t max_size)
 {
 	Cache *cache;
